ui: add function for user accept and reject pending devices
The user can access the pending devices in the notification list,when in the
list of pending devices, the user can accept or decline.

When the user is in /devices/pending and accepts a device in the notification
list, a logic was created in which the numberPage and the numberPerPage are
stored in devices.js. Which is used by an action when called on the
notification component to update devices in /devices/pending. The data is in
separate files and one component needs to update the other when the user
performs an acceptance or rejection action.